Fixed various Win32 PATH related build problems:
The main problem fixed was that PATH was not being setup at the proper time to find the target compiler, requiring that you instead put the compiler on the PATH so it could be found. The makefiles already had the logic to make it so this was not necessary, but it was broken, mainly because WIN2POSIX was being called on some paths before WIN2POSIX was setup. To fix this, tools/tools.gmk is now included before including any of the platform makefiles rather than after.

Another problem was not properly deferring the setting of PATH until after refereces were made to it, but also setting it before it was used by rules.mk makefiles. Setting it in the GNUmakefile after including top.mk does not work because it is too late, but setting it in the GNUmakefile before including top.mk is too early because of the WIN2POSIX references. Basically it needs to be done from a defs.mk makefile, either in the build directory, or in a makefile included from the defs.mk makefile in the build directory.

In win32/host_defs.mk, fixed the echoing of TARGET_CC_VERSION, which would do nothing unless TARGET_CC was on the path. This was fixed by changing the command from "$(TARGET_CC)" to "PATH=$(PATH); $(TARGET_CC)".

In win32/rules.mk, fix CVM_COMPILER_INCOMPATIBLE flag setting in win32/rule.mk. It is not set because it happens after share/rules.mk already sets it false, therefore it did not detect incompatible compilers. It also suffered from PATH problems, which are fixed by setting PATH in the $(TARGET_CC) command.

Reviewed by Dean.